+++ ports/nrf/drivers/rng.c
@@ -3,6 +3,7 @@
  *
  * The MIT License (MIT)
  *
+ * Copyright (c) 2017-2018 Glenn Ruben Bakke
  * Copyright (c) 2018 Ayke van Laethem
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -24,4 +25,55 @@
  * THE SOFTWARE.
  */
 
-uint32_t machine_rng_generate_random_word(void);
+#include "py/mpconfig.h"
+
+#if MICROPY_PY_RANDOM_HW_RNG
+
+#include "nrf_rng.h"
+#include "drivers/rng.h"
+
+#if BLUETOOTH_SD
+#include "nrf_soc.h"
+#include "ble_drv.h"
+#define BLUETOOTH_STACK_ENABLED() (ble_drv_stack_enabled())
+#endif
+
+static inline uint32_t generate_hw_random(void) {
+    uint32_t retval = 0;
+    uint8_t * p_retval = (uint8_t *)&retval;
+
+    nrf_rng_event_clear(NRF_RNG_EVENT_VALRDY);
+    nrf_rng_task_trigger(NRF_RNG_TASK_START);
+
+    for (uint16_t i = 0; i < 4; i++) {
+        while (!nrf_rng_event_get(NRF_RNG_EVENT_VALRDY)) {
+            ;
+        }
+
+        nrf_rng_event_clear(NRF_RNG_EVENT_VALRDY);
+        p_retval[i] = nrf_rng_random_value_get();
+    }
+
+    nrf_rng_task_trigger(NRF_RNG_TASK_STOP);
+
+    return retval;
+}
+
+uint32_t rng_generate_random_word(void) {
+
+#if BLUETOOTH_SD
+    if (BLUETOOTH_STACK_ENABLED() == 1) {
+        uint32_t retval = 0;
+        uint32_t status;
+        do {
+            status = sd_rand_application_vector_get((uint8_t *)&retval, 4); // Extract 4 bytes
+        } while (status != 0);
+
+        return retval;
+    }
+#endif
+
+    return generate_hw_random();
+}
+
+#endif // MICROPY_PY_RANDOM_HW_RNG
