+++ ports/nrf/modules/uos/microbitfs.c
@@ -31,7 +31,7 @@
 
 #include "microbitfs.h"
 #include "drivers/flash.h"
-#include "modrandom.h"
+#include "drivers/rng.h"
 #include "py/obj.h"
 #include "py/stream.h"
 #include "py/runtime.h"
@@ -175,7 +175,7 @@ STATIC void init_limits(void) {
 }
 
 STATIC void randomise_start_index(void) {
-    start_index = machine_rng_generate_random_word() % chunks_in_file_system + 1;
+    start_index = rng_generate_random_word() % chunks_in_file_system + 1;
 }
 
 void microbit_filesystem_init(void) {
